+++ cmake/ExternalDependencies.cmake
@@ -43,7 +43,7 @@ set(QDMI_MINIMUM_VERSION 1.4.0
         CACHE STRING "Minimum QDMI version")
 set(QDMI_VERSION 1.4.0
         CACHE STRING "QDMI version")
-set(QDMI_REV "5f572b249f596dabe3750b34deff470f33eb917f"
+set(QDMI_REV "b996400e3e85455878e49c0615e2d5b8d514cab6"
         CACHE STRING "QDMI identifier (tag, branch or commit hash)")
 set(QDMI_REPO_OWNER "Munich-Quantum-Software-Stack"
         CACHE STRING "QDMI repository owner (change when using a fork)")